backend/requirements.txt
# Core
Django==5.1.6
djangorestframework==3.15.0
psycopg2-binary==2.9.9
pillow==10.3.0
python-decouple==3.8

# Admin UI
django-admin-interface==0.30.0

# Security
django-cors-headers==4.3.1
django-ratelimit==4.1.0

# Utilities
django-filter==23.5
django-encrypted-fields==1.1.2
gunicorn==21.2.0

# Email
sendgrid==6.10.0

# For future phases
# djangorestframework-simplejwt==5.3.1
# channels==4.0.0
# daphne==4.0.0
# django-storages==1.14.2
# boto3==1.34.14 